+++ tests/test_settings.py
@@ -2540,7 +2540,7 @@ def main() -> None:
     defaults = {}
     setting_ids = set()
 
-    assert data.version == "2.6.1"
+    assert data.version == "2.6.3"
     assert (
         localization["quick_look_card_integration_group"]["en"]
         == "Mod Integration: Quick Look Card"